feat: extract unix persmission in a trait
